docs: Added Access Control
Added Access Control to the Job Controller limiting the actions a user can do to create, edit, delete. User must be logged in as themselves in order to create, edit, delete.

Adding Access Control to the Job Controller makes it so any user can not just edit or delete a specific users posting.